cloud/api/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
supabase>=2.4.0
python-jose[cryptography]>=3.3.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
httpx>=0.27.0
